+++ src/main/java/com/poratu/idea/plugins/tomcat/runner/TomcatDebugger.java
@@ -1,10 +1,16 @@
 package com.poratu.idea.plugins.tomcat.runner;
 
 import com.intellij.debugger.impl.GenericDebuggerRunner;
+import com.intellij.execution.ExecutionException;
+import com.intellij.execution.configurations.RemoteConnection;
 import com.intellij.execution.configurations.RunProfile;
+import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.executors.DefaultDebugExecutor;
+import com.intellij.execution.runners.ExecutionEnvironment;
+import com.intellij.execution.ui.RunContentDescriptor;
 import com.poratu.idea.plugins.tomcat.conf.TomcatRunConfiguration;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Author : zengkid
@@ -24,5 +30,14 @@ public class TomcatDebugger extends GenericDebuggerRunner {
     public boolean canRun(@NotNull String executorId, @NotNull RunProfile profile) {
         return (DefaultDebugExecutor.EXECUTOR_ID.equals(executorId) && profile instanceof TomcatRunConfiguration);
     }
+
+    /*
+    @Nullable
+    @Override
+    protected RunContentDescriptor createContentDescriptor(@NotNull RunProfileState state, @NotNull ExecutionEnvironment environment) throws ExecutionException {
+        // Create a remote connection to the debug port (8081 as defined in JAVA_DEBUG_OPTIONS)
+        RemoteConnection connection = new RemoteConnection(true, "localhost", "5005", false);
+        return attachVirtualMachine(state, environment, connection, true);
+    } */
 
 }