+++ src/main/java/com/poratu/idea/plugins/tomcat/setting/TomcatServersConfigurable.java
@@ -2,15 +2,23 @@ package com.poratu.idea.plugins.tomcat.setting;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.project.DumbAwareAction;
+import com.intellij.openapi.roots.libraries.Library;
+import com.intellij.openapi.roots.libraries.LibraryTable;
+import com.intellij.openapi.roots.libraries.LibraryTablesRegistrar;
 import com.intellij.openapi.ui.MasterDetailsComponent;
+import com.intellij.openapi.vfs.JarFileSystem;
+import com.intellij.openapi.vfs.LocalFileSystem;
+import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.CommonActionsPanel;
 import com.intellij.util.IconUtil;
 import com.poratu.idea.plugins.tomcat.utils.PluginUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,7 +28,6 @@ import java.util.List;
  * Time   : 00:14
  */
 public class TomcatServersConfigurable extends MasterDetailsComponent {
-
     @Override
     public String getDisplayName() {
         return "Tomcat Server";
@@ -75,8 +82,65 @@ public class TomcatServersConfigurable extends MasterDetailsComponent {
 
         for (int i = 0; i < myRoot.getChildCount(); i++) {
             TomcatInfoConfigurable configurable = (TomcatInfoConfigurable) ((MyNode) myRoot.getChildAt(i)).getConfigurable();
-            tomcatInfos.add(configurable.getEditableObject());
+            TomcatInfo tomcatInfo = configurable.getEditableObject();
+            tomcatInfos.add(tomcatInfo);
+
+            // Create or update global library for this Tomcat
+            createOrUpdateTomcatLibrary(tomcatInfo);
         }
+    }
+
+    /**
+     * Creates or updates a global library for the specified Tomcat server
+     * @param tomcatInfo The Tomcat server information
+     */
+    private void createOrUpdateTomcatLibrary(TomcatInfo tomcatInfo) {
+        ApplicationManager.getApplication().runWriteAction(() -> {
+            String libraryName = tomcatInfo.getName();
+            LibraryTable libraryTable = LibraryTablesRegistrar.getInstance().getLibraryTable();
+
+            // Remove existing library if it exists
+            Library existingLibrary = libraryTable.getLibraryByName(libraryName);
+            if (existingLibrary != null) {
+                libraryTable.removeLibrary(existingLibrary);
+            }
+
+            // Create new library
+            Library library = libraryTable.createLibrary(libraryName);
+            Library.ModifiableModel libraryModel = library.getModifiableModel();
+
+            // Add all JARs from Tomcat's lib directory
+            File libDir = new File(tomcatInfo.getPath(), "lib");
+            if (libDir.exists() && libDir.isDirectory()) {
+                File[] jarFiles = libDir.listFiles((dir, name) -> name.toLowerCase().endsWith(".jar"));
+                if (jarFiles != null) {
+                    for (File jarFile : jarFiles) {
+                        VirtualFile jarVirtualFile = LocalFileSystem.getInstance().findFileByIoFile(jarFile);
+                        if (jarVirtualFile != null) {
+                            VirtualFile jarRoot = JarFileSystem.getInstance().getJarRootForLocalFile(jarVirtualFile);
+                            if (jarRoot != null) {
+                                libraryModel.addRoot(jarRoot, com.intellij.openapi.roots.OrderRootType.CLASSES);
+                            }
+                        }
+                    }
+                }
+            }
+
+            // Add servlet-api.jar from Tomcat's lib directory
+            File servletApiJar = new File(libDir, "servlet-api.jar");
+            if (servletApiJar.exists()) {
+                VirtualFile jarVirtualFile = LocalFileSystem.getInstance().findFileByIoFile(servletApiJar);
+                if (jarVirtualFile != null) {
+                    VirtualFile jarRoot = JarFileSystem.getInstance().getJarRootForLocalFile(jarVirtualFile);
+                    if (jarRoot != null) {
+                        libraryModel.addRoot(jarRoot, com.intellij.openapi.roots.OrderRootType.CLASSES);
+                    }
+                }
+            }
+
+            // Commit changes
+            libraryModel.commit();
+        });
     }
 
     @Override